+++ src/Transaction.java
@@ -1,8 +1,10 @@
+import java.math.BigDecimal;
+
 public class Transaction{
       public char type_;
       public String from_;
       public String to_;
-      public Integer value_;
+      public BigDecimal value_;
 
       public char getType(){
           return this.type_;
